+++ src/views/Main/index.js
@@ -469,7 +469,7 @@ export default function Main(props) {
 			<DebugModal
 				open={$processDebug.open}
 				onClose={handleProcessDebug}
-				title={<Trans>Process debug report</Trans>}
+				title={<Trans>Process report</Trans>}
 				data={$processDebug.data}
 				onHelp={handleHelp('process-report')}
 			/>